+++ pleroma_bot/_processing.py
@@ -80,13 +80,15 @@ def process_tweets(self, tweets_to_post):
     all_media = []
     for tweet in tweets_to_post["data"]:
         self.posts_ids[self.pleroma_base_url].update({tweet["id"]: ''})
+        tweet["reply_id"] = None
+        tweet["retweet_id"] = None
         # get reply ids
-        if "in_reply_to_user_id" in tweet.keys():
+        if "referenced_tweets" in tweet.keys():
             for reference in tweet["referenced_tweets"]:
                 if reference["type"] == "replied_to":
                     tweet["reply_id"] = reference["id"]
-        else:
-            tweet["reply_id"] = None
+                if reference["type"] == "retweeted":
+                    tweet["retweet_id"] = reference["id"]
         media = []
         logger.debug(tweet["id"])
         # Get full text from RT or quoted tweet