+++ pleroma_bot/tests/test_logger.py
@@ -126,7 +126,7 @@ def test_post_pleroma_media_logger(rootdir, sample_users, caplog):
                 mock.post(media_url, status_code=413)
                 with caplog.at_level(logging.ERROR):
                     sample_user_obj.post_pleroma(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
                 assert 'Exception occurred' in caplog.text
                 assert 'Media size too large' in caplog.text
@@ -136,7 +136,7 @@ def test_post_pleroma_media_logger(rootdir, sample_users, caplog):
                         requests.exceptions.HTTPError
                 ) as error_info:
                     sample_user_obj.post_pleroma(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
                 exception_value = (
                     f"500 Server Error: None for url: {media_url}"
@@ -150,7 +150,7 @@ def test_post_pleroma_media_logger(rootdir, sample_users, caplog):
                 mock.post(media_url, status_code=422, json=error_json)
                 with caplog.at_level(logging.ERROR):
                     sample_user_obj.post_pleroma(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
 
                 assert error_json["error"] in caplog.text
@@ -192,7 +192,7 @@ def test_post_misskey_media_logger(rootdir, sample_users, caplog):
                 mock.post(media_url, json={}, status_code=413)
                 with caplog.at_level(logging.ERROR):
                     sample_user_obj.post_misskey(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
                 assert 'Exception occurred' in caplog.text
                 assert 'Media size too large' in caplog.text
@@ -202,7 +202,7 @@ def test_post_misskey_media_logger(rootdir, sample_users, caplog):
                         requests.exceptions.HTTPError
                 ) as error_info:
                     sample_user_obj.post_misskey(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
                 exception_value = (
                     f"500 Server Error: None for url: {media_url}"
@@ -216,7 +216,7 @@ def test_post_misskey_media_logger(rootdir, sample_users, caplog):
                 mock.post(media_url, status_code=422, json=error_json)
                 with caplog.at_level(logging.ERROR):
                     sample_user_obj.post_misskey(
-                        (test_user.pinned, "", "", None), None, False
+                        (test_user.pinned, "", "", None, None), None, False
                     )
 
                 assert error_json["error"] in caplog.text